deep_continuation/plots/requirements.txt
argh==0.26.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
deep-continuation==0.2
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.5
gitpython==3.1.8
idna==2.10
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.19.1
pathtools==0.1.2
pip==19.1.1
promise==2.3
protobuf==3.13.0
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyyaml==5.3.1
requests==2.24.0
scipy==1.4.1
sentry-sdk==0.18.0
setuptools==41.0.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.3
torch==1.6.0
urllib3==1.25.10
wandb==0.10.3
watchdog==0.9.0
wheel==0.33.4